Add ignoring of worktree directory.
The more I work with git worktree's the more I like them, the issue is
that when I'm searching or looking for files and I'm not in my worktree
directory (the one I use to house them all) I don't want to look
through them.

Adding them to ignore for ag and ctrlp.