Promote vim-polyglot
Having this syntax module installed is amazing. Having proper syntax
highlighting for pretty much anything without additional installs is
extremely helpful.

Because of which, removing the other syntax modules as vim-polyglot
already includes them, except for vim-perl, the version with
vim-polyglot doesn't have the same options which really are required.