Move to Makefile based installation
Getting tired of the .dotfiles repository being a mess of installations
and configration files, this is the begining of categorizing dot files
by the applications/service to which they belong, and using a Makefile
to do the necessary linking/backing up.

First on the list is neovim, it includes installing vim-plug via curl,
which may move later to an Ansible role/playbook in cases where the
installation target does not have internet access.