Change fish configuration to Makefile
Continuing in the series of making services installable by Makefile,
moving fish and adding the Makefile. The Makefile does automatically
install the fisher plugin via curl, but as I only run fishshell on
workstations I find this more acceptable. It may change however to be
part of an Ansible role in the future too.