Change tmux configuration installation
Due to recent changes in tmux with regards to names of configuration
options, and that the macOS version of tmux from homebrew installs
version `2.9a` (the `a` being the part that breaks the use of `bc` to
determine version differences). Create a new installation process
following the Makefiles used earlier to link the appropriate tmux
configuration.

The added benefit is that the tpm plugin is no longer in version control
as well.