Add bass for sourcing bash files in fish
Some times people write scripts that set environment variables for the
cli environment. Because fish uses a different syntax setting them can
be troublesome. This is where bass comes in. It allows you to point to
the file that should be sourced and imports directly into the fish
shell, no need for manual conversion.