Migrate to using an opt object
It seems there is an issue with setting options that some are buffer and
global and some are window and global, but setting buffer or window
alone is not enough to have them be applied to new  buffers. There's
work being done to fix that but until the PR is merged, this workaround
is in place.

https://www.reddit.com/r/neovim/comments/l6a8z8/is_it_just_me_or_some_options_dont_work_in_lua_yet/