conf(neovim): Separate telescope file browser key mappings
Sometimes I want to browse files in the same directory as the current
buffer and sometimes I want to browse from where I've started neovim.
Constantly hitting enter on .. in the former case is annoying.

Set key mappings for fb to browse from where neovim started, and fl
to browse the directory of the current buffer.