chore(gitconfig): clean up unused options
With the move for work to be using GitHub, I needed to turn off the
check for resolvable email addresses on my account, so now setting the
default email to the noreply address. I'll need to configure real email
addresses appropriately going forward.

There was a bunch of cruft in the config that's never used, removing
that too.